fix: 인기 상권 목록 조회 api 요청 시 타입 에러 해결 #45

+++ src/main/java/com/example/smaap/domain/region/service/NeighborhoodService.java
@@ -149,7 +149,7 @@ public class NeighborhoodService {
                 .select(Projections.constructor(NeighborhoodCountDTO.Response.class,
                         neighborhood.id,
                         neighborhood.name,
-                        sumAllHours.sum()))
+                        sumAllHours.sum().longValue()))
                 .from(neighborhood)
                 .leftJoin(population)
                 .on(population.neighborhood.id.eq(neighborhood.id))
@@ -193,7 +193,7 @@ public class NeighborhoodService {
                 .select(Projections.constructor(NeighborhoodCountDTO.Response.class,
                         neighborhood.id,
                         neighborhood.name,
-                        sumAllHours.sum()))
+                        sumAllHours.sum().longValue()))
                 .from(neighborhood)
                 .leftJoin(population)
                 .on(population.neighborhood.id.eq(neighborhood.id))
